refactor out shared commands
