+++ torchci/components/runners/StatusChip.tsx
@@ -1,0 +1,27 @@
+import { Chip } from "@mui/material";
+import { RunnerData } from "lib/runnerUtils";
+
+export function StatusChip({ runner }: { runner: RunnerData }) {
+  let color: "success" | "warning" | "default";
+  let label: string;
+
+  if (runner.status === "offline") {
+    color = "default";
+    label = "offline";
+  } else if (runner.busy) {
+    color = "warning";
+    label = "busy";
+  } else {
+    color = "success";
+    label = "idle";
+  }
+
+  return (
+    <Chip
+      label={label}
+      color={color}
+      size="small"
+      sx={{ minWidth: 80, fontWeight: "bold" }}
+    />
+  );
+}
